3.5 germline assay : version before code optimization (update inheritance)
