+++ HMVV/src/hmvv/gui/mutationlist/tables/CommonTable.java
@@ -245,10 +245,10 @@ public abstract class CommonTable extends JTable{
 		Mutation mutation = getSelectedMutation();		
 		String gene = mutation.getGene();
 		
-		ArrayList<GeneAnnotation> geneAnnotationHistory = DatabaseCommands.getGeneAnnotationHistory(gene,Configurations.MUTATION_TYPE.SOMATIC);
+		ArrayList<GeneAnnotation> geneAnnotationHistory = DatabaseCommands.getGeneAnnotationHistory(gene,mutation.getMutationType());
 
 
-		AnnotationFrame editAnnotation = new AnnotationFrame(mutation, geneAnnotationHistory, this, parent);
+		AnnotationFrame editAnnotation = new AnnotationFrame(mutation, geneAnnotationHistory, parent);
 		editAnnotation.setVisible(true);
 		this.setCursor(Cursor.getDefaultCursor());
 	}
