+++ HMVV/src/hmvv/gui/mutationlist/tables/GermlineSNPEFFTable.java
@@ -20,7 +20,7 @@ public class GermlineSNPEFFTable extends CommonTableGermline{
 
         @Override
         protected HMVVTableColumn[] constructCustomColumns(){
-            return HMVVTableColumn.getCustomColumnArray(model.getColumnCount(), 2,3,4,10,11,18);
+            return HMVVTableColumn.getCustomColumnArray(model.getColumnCount(), 2,3,4,10,11,19);
         }
 
         @Override
@@ -43,7 +43,7 @@ public class GermlineSNPEFFTable extends CommonTableGermline{
                 searchGoogleForDNAChange();
             } else if (column == 11) {
                 searchGoogleForProteinChange();
-            } else if (column == 18) {
+            } else if (column == 19) {
                 handleAnnotationClick();
             }
         }
